Get active project data
