+++ pages/index.js
@@ -4,66 +4,47 @@ import styles from '../styles/Home.module.css'
 
 export default function Home() {
   return (
-    <div className={styles.container}>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-
-      <main className={styles.main}>
-        <h1 className={styles.title}>
-          Welcome to <a href="https://nextjs.org">Next.js!</a>
-        </h1>
-
-        <p className={styles.description}>
-          Get started by editing{' '}
-          <code className={styles.code}>pages/index.js</code>
-        </p>
-
-        <div className={styles.grid}>
-          <a href="https://nextjs.org/docs" className={styles.card}>
-            <h2>Documentation &rarr;</h2>
-            <p>Find in-depth information about Next.js features and API.</p>
-          </a>
-
-          <a href="https://nextjs.org/learn" className={styles.card}>
-            <h2>Learn &rarr;</h2>
-            <p>Learn about Next.js in an interactive course with quizzes!</p>
-          </a>
-
-          <a
-            href="https://github.com/vercel/next.js/tree/canary/examples"
-            className={styles.card}
-          >
-            <h2>Examples &rarr;</h2>
-            <p>Discover and deploy boilerplate example Next.js projects.</p>
-          </a>
-
-          <a
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-          >
-            <h2>Deploy &rarr;</h2>
-            <p>
-              Instantly deploy your Next.js site to a public URL with Vercel.
-            </p>
-          </a>
-        </div>
-      </main>
-
-      <footer className={styles.footer}>
-        <a
-          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Powered by{' '}
-          <span className={styles.logo}>
-            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
-          </span>
-        </a>
-      </footer>
-    </div>
+    <>
+    </>
   )
 }
+
+
+export async function getStaticProps() {
+    
+  const notion = new Client({auth: process.env.NOTION_API_KEY});
+  const response = await notion.search({
+      filter: {
+          property: "object",
+          value: "database"
+      }
+  });
+  
+  const ProjectID = response.results[0].id;
+  const PeopleID = response.results[1].id;
+  const TimereportID = response.results[2].id;
+
+  const People = await notion.databases.query({
+      database_id: PeopleID,
+  });
+  const Timereports = await notion.databases.query({
+      database_id: TimereportID,
+  });
+  const ActiveProjects = await notion.databases.query({
+      database_id: ProjectID,
+      filter:{
+        property: "Status",
+        select: {
+          equals: "Active"
+        }
+      } 
+  });
+
+  return {
+      props: {
+          projects: ActiveProjects.results 
+      }
+  }
+}
+
+
