--- pages/test.js
@@ -1,84 +1,0 @@
-import Head from 'next/head'
-import Image from 'next/image'
-import styles from '../styles/Home.module.css'
-import { Client } from '@notionhq/client'
-
-export default function Test({People, Timereports, Projects}) {
-    console.log(People);
-  return (
-    <div className={styles.container}>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-
-      <main className={styles.main}>
-        <h1 className={styles.title}>
-          Welcome to <a href="https://nextjs.org">Next.js!</a>
-        </h1>
-
-        <p className={styles.description}>
-          Get started by editing{' '}
-          <code className={styles.code}>pages/index.js</code>
-        </p>
-        <label for="first">Name:</label>
-        <input type="text" id="first" name="first" />
-
-        <div className={styles.grid}>
-          <a href="https://nextjs.org/docs" className={styles.card}>
-            <h2>Documentation &rarr;</h2>
-            <p>Find in-depth information about Next.js features and API.</p>
-          </a>
-          <form action="/send-data-here" method="post">
-        
-        </form>
-         
-        </div>
-      </main>
-      <label for="first">First name:</label>
-        <input type="text" id="first" name="first" />
-        <button type="submit">Submit</button>
-    </div>
-  )
-}
-
-function form() {
-
-
-
-    
-}
-
-//https://www.notion.so/8d2454d4cb1540dda8936973268957b7?v=d9743ee5220f42789b4ad88dc1af9264
-export async function getStaticProps() {
-    const notion = new Client({auth: process.env.NOTION_API_KEY});
-    const response = await notion.search({
-        filter: {
-            property: "object",
-            value: "database"
-        }
-    });
-    
-    const PeopleID = response.results[0].id;
-    const TimereportID = response.results[1].id;
-    const ProjectID = response.results[2].id;
-
-    const People = await notion.databases.query({
-        database_id: PeopleID,
-    });
-    const Timereports = await notion.databases.query({
-        database_id: TimereportID,
-    });
-    const Projects = await notion.databases.query({
-        database_id: ProjectID,
-    });
-
-    return {
-        props: {
-            People,
-            Timereports,
-            Projects
-        }
-    }
-}
